refactor(frontend): PR comments
