+++ frontend/src/component/FlashMessage.styl
@@ -38,7 +38,7 @@
         &__paragraph
           padding-right 20px
 
-#connection_error
+.connection_error
   z-index 2000
 
 @media (max-width: max-sm)
