+++ frontend/src/reducer/root.js
@@ -12,7 +12,7 @@ import contentType from './contentType.js'
 import timezone from './timezone.js'
 import system from './system.js'
 import searchResult from './searchResult.js'
-import tlm from './tlmManager.js'
+import tlm from './tlm.js'
 
 const rootReducer = combineReducers({
   lang,
