+++ frontend/src/reducer/tlm.js
@@ -1,6 +1,12 @@
+import { LIVE_MESSAGE_STATUS } from '../util/LiveMessageManager.js'
 import { SET, TLM_MANAGER, TLM_MANAGER_STATUS } from '../action-creator.sync.js'
 
-export default function tlm (state = {}, action) {
+const defaultTlmState = {
+  manager: null,
+  status: LIVE_MESSAGE_STATUS.CLOSED
+}
+
+export default function tlm (state = defaultTlmState, action) {
   switch (action.type) {
     case `${SET}/${TLM_MANAGER}`:
       return { ...state, manager: action.TLMManager }
